# concierge/api/openapi.yaml
openapi: 3.0.3
info:
  title: concierge — Basis SSO admission & control plane API
  version: "1.0.0"
  description: >
    Wire-compatible Go rewrite of the Basis SSO admission broker
    (the former Basis SSO broker). Endpoint paths, JSON field names, and
    status codes are load-bearing: they are consumed by unmodified C# Unity
    clients and the unmodified C# UDP game server. See
    docs/concierge/research-sso-broker.md and docs/concierge/design.md for
    the source specification. HTML/plain-text/raw-file responses (enroll and
    join pages, /admin/client-config/{serverId}) are documented here for
    completeness but are written directly by the handlers rather than
    generated, since oapi-codegen's std-http-server mode only generates
    routing and parameter binding, not response encoding. Managed WebGL
    GameServers may terminate TLS using the read-only Kubernetes Secret
    configured by Broker.Kubernetes; that deployment-only setting is not
    mutable through this API.
paths:
  /client-config/{serverId}:
    get:
      operationId: getClientConfig
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Stored client configuration with clientSecret fields stripped.
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Server unknown or no client-config file stored.
...
  /admin/organization:
    get:
      operationId: getOrganization
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationOptions"
        "401":
          description: Unauthorized.
    put:
      operationId: putOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationOptions"
      responses:
        "204":
          description: Saved.
        "400":
          description: Structurally invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
  /admin/meetings:
    get:
      operationId: listMeetings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetingView"
        "401":
          description: Unauthorized.
    post:
      operationId: createMeeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingRequest"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingView"
        "400":
          description: Invalid title/id/host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
        "409":
          description: Organization SSO settings incomplete.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Kubernetes GameServer provisioning failed, including invalid TLS Secret configuration.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "501":
          description: Manual meeting creation not enabled.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /admin/meetings/{meetingId}:
    delete:
      operationId: deleteMeeting
      parameters:
        - $ref: "#/components/parameters/MeetingId"
      responses:
        "204":
          description: Deleted.
        "401":
          description: Unauthorized.
        "404":
          description: Unknown meeting.
  /admin/meetings/{meetingId}/invitations:
    post:
      operationId: createInvitation
      parameters:
        - $ref: "#/components/parameters/MeetingId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "401":
          description: Unauthorized.
        "404":
          description: Unknown meeting.
  /admin/servers/{serverId}:
    put:
      operationId: putAdminServer
      parameters:
        - $ref: "#/components/parameters/ServerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminServerWrite"
      responses:
        "204":
          description: Upserted.
        "400":
          description: Structurally invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
    delete:
      operationId: deleteAdminServer
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "204":
          description: Deleted.
        "401":
          description: Unauthorized.
        "404":
          description: Unknown server.
  /admin/client-config-template/{serverId}:
    get:
      operationId: getClientConfigTemplate
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Pretty-printed canonical client configuration, including clientSecret.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientConfiguration"
        "401":
          description: Unauthorized.
        "404":
          description: Unknown server.
        "503":
          description: Transport public key not available.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /admin/client-config/{serverId}:
    get:
      operationId: getAdminClientConfig
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Raw stored client-config file bytes.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized.
        "404":
          description: Unknown server or no file stored.
    put:
      operationId: putAdminClientConfig
      parameters:
        - $ref: "#/components/parameters/ServerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: Stored.
        "400":
          description: Body too large or not a JSON object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
        "404":
          description: Unknown server.
        "503":
          description: Client config directory not configured.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  parameters:
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
...
    MeetingId:
      name: meetingId
      in: path
      required: true
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ProblemDetails:
      type: object
      description: RFC 7807 application/problem+json envelope, matching ASP.NET Core's Results.Problem default shape.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    ProviderOptions:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        issuer:
          type: string
        audience:
          type: string
        clientSecret:
          type: string
        webClientId:
          type: string
        webClientSecret:
          type: string
        tokenEndpoint:
          type: string
        jwksUri:
          type: string
        allowedHostedDomains:
          type: array
          items:
            type: string
        allowedGroups:
          type: array
          items:
            type: string
    OrganizationOptions:
      type: object
      properties:
        displayName:
          type: string
        defaultProviderId:
          type: string
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderOptions"
...
    AdminServerWrite:
      type: object
      description: Body for PUT /admin/servers/{serverId} (BrokerServerOptions equivalent).
      properties:
        id:
          type: string
        ticketSigningKeyEnvironmentVariable:
          type: string
        transportPublicKeyEnvironmentVariable:
          type: string
        ticketSigningKey:
          type: string
        transportPublicKey:
          type: string
        webSocketUri:
          type: string
        serverInfoUri:
          type: string
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderOptions"
    MeetingView:
      type: object
      required: [id, title, status, statusDetail, host, port, createdAt, updatedAt, joinUrl, invitationReady]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        statusDetail:
          type: string
        host:
          type: string
        port:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        joinUrl:
          type: string
        webJoinUrl:
          type: string
          description: >-
            Browser (WebGL) join URL, omitted when no allowed web origin can
            serve the web client.
        invitationReady:
          type: boolean
        webSocketUri:
          type: string
        serverInfoUri:
          type: string
    CreateMeetingRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        id:
          type: string
        host:
          type: string
        port:
          type: integer
        password:
          type: string
        webSocketUri:
          type: string
        serverInfoUri:
          type: string
    InvitationResponse:
      type: object
      required: [url, meetingId]
      properties:
        url:
          type: string
        meetingId:
          type: string
...
    ServerTransportConfig:
      type: object
      properties:
        serverPublicKey:
          type: string
        admissionEndpoint:
          type: string
        allowUntrustedLoopbackCertificate:
          type: boolean
        webSocketUri:
          type: string
        serverInfoUri:
          type: string
    ClaimRule:
      type: object
      properties:
        claim:
          type: string
        values:
          type: array
          items:
            type: string
    AccessConfig:
      type: object
      properties:
        allowedGroups:
          type: array
          items:
            type: string
        allowedClaims:
          type: array
          items:
            $ref: "#/components/schemas/ClaimRule"
    ClientProviderConfig:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        issuer:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        tokenEndpoint:
          type: string
        extraAuthParams:
          type: object
          additionalProperties:
            type: string
        scopes:
          type: array
          items:
            type: string
        displayNameClaims:
          type: array
          items:
            type: string
        access:
          $ref: "#/components/schemas/AccessConfig"
    RedirectConfig:
      type: object
      properties:
        mode:
          type: string
        host:
          type: string
        port:
          type: integer
        path:
          type: string
    EnforcementConfig:
      type: object
      properties:
        allowOfflineWithinTokenValidity:
          type: boolean
...
    ClientConfiguration:
      type: object
      properties:
        defaultProviderId:
          type: string
        serverTransport:
          $ref: "#/components/schemas/ServerTransportConfig"
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ClientProviderConfig"
        redirect:
          $ref: "#/components/schemas/RedirectConfig"
        enforcement:
          $ref: "#/components/schemas/EnforcementConfig"
...